refactor: Merge stats db and server into one exe
* Logical to have both of them in one binary. It reduces maintainence and operation costs

* Spawn db update and server in separate go routines
