--- cmd/batchjob_stats_db/batchjob_stats_db_test.go
@@ -1,32 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"os"
-	"os/exec"
-	"path/filepath"
-	"testing"
-)
-
-var (
-	binary, _ = filepath.Abs("bin/batchjob_stats_db")
-)
-
-func TestBatchjobStatExecutable(t *testing.T) {
-	if _, err := os.Stat(binary); err != nil {
-		t.Skipf("batchjob_stats_db binary not available, try to run `make build` first: %s", err)
-	}
-	tmpDir := t.TempDir()
-
-	jobstats := exec.Command(binary, "--data.path", tmpDir)
-	if err := runCommandAndTests(jobstats); err != nil {
-		t.Error(err)
-	}
-}
-
-func runCommandAndTests(cmd *exec.Cmd) error {
-	if err := cmd.Start(); err != nil {
-		return fmt.Errorf("failed to start command: %s", err)
-	}
-	return nil
-}
